System automatically submitted

+++ os/src/syscall/process.rs
@@ -1029,7 +1029,7 @@ pub fn sys_getrusage(who: isize, usage: *mut Rusage) -> isize {
 pub fn sys_mycall() -> isize {
     let name: &'static str = "My name is Zhang Jiawen";
     for c in name.chars() {
-        console_putchar(c);
+        console_putchar(c as usize);
     }
     0
 }
